fix: convertTask test
